handle warning changes

+++ projects/bi-modules/src/lib/components/BI-Grid/bi-grid.component.ts
@@ -1,5 +1,5 @@
-import { AfterViewInit, ChangeDetectorRef, Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
-import { CellClickEvent, CellCloseEvent, CreateFormGroupArgs, GridComponent, GridItem, PageChangeEvent } from '@progress/kendo-angular-grid';
+import { ChangeDetectorRef, Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { CellClickEvent, CellCloseEvent, CreateFormGroupArgs, GridComponent, GridItem } from '@progress/kendo-angular-grid';
 import { State, toODataString } from "@progress/kendo-data-query";
 import { AlertService } from '@full-fledged/alerts';
 import { FormBuilder, FormGroup } from '@angular/forms';
@@ -8,12 +8,14 @@ import { IDataSource } from 'bi-interfaces/lib/interfaces/IDataSource';
 import { IGrid } from 'bi-interfaces/lib/interfaces/IGrid';
 import { Subject } from 'rxjs';
 import { take } from 'rxjs/operators';
+import Swal from 'sweetalert2';
+
 @Component({
 	selector: 'BI-Grid',
 	templateUrl: './bi-grid.component.html',
 	styleUrls: ['./bi-grid.component.scss']
 })
-export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
+export class BIGridComponent implements IGrid, OnInit {
 	@Input() public DataService!: IDataSource;
 	@Input() Columns!: IColumns[];
 	@Input() GridName!: string;
@@ -21,14 +23,16 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 	@Output() CellClick = new EventEmitter<CellClickEvent>();
 	@ViewChild("Grid") Mygrid!: GridComponent;
 	StopSave!: Subject<boolean>;
+	StopDelete!: Subject<boolean>;
 	CreatedItemArray: Array<Object> = [];
 	UpdatedItemArray: Array<Object> = [];
 	form: any = {};
 	CurrentSelectRow!: FormGroup;
 	GridData!: any;
 	state: State = { skip: 0, take: 10 };
-	rowIndex!: number;
+	rowIndex!: any;
 	dataItem!: any;
+	dataItemReset!: any;
 	data: any;
 	newForm: any = {};
 
@@ -46,11 +50,10 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 		this.createFormGroup = this.createFormGroup.bind(this);
 		this.StopSave = new Subject<boolean>();
 		this.StopSave.next(true);
+		this.StopDelete = new Subject<boolean>();
+		this.StopDelete.next(true);
 	}
 
-	ngAfterViewInit(): void {
-
-	}
 	handleFormGroup() {
 		this.Columns.forEach(res => {
 			if (this.form.hasOwnProperty(res.Name)) this.form[res.Name] = [{ value: res.DefaultValue, disabled: !res.IsEditable }, res.Validators]
@@ -66,18 +69,20 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 	createFormGroup(args: CreateFormGroupArgs | any): FormGroup {
 		this.rowIndex = args.rowIndex;
 		const item = args.dataItem;
+		this.dataItemReset = { ...args.dataItem };
 		this.CurrentSelectRow.patchValue(item);
 		return this.CurrentSelectRow;
 	}
 
 	GetGridData() {
 		this.GridData = this.DataService;
-		this.GridData.subscribe((res: any) => {
-			this.data = res;
-		});
+		this.GridData.subscribe((res: any) => this.data = res);
 	}
 
-	BeforeAction(): void {
+	BeforeActionSave(): void {
+	}
+
+	BeforeActionDelete(): void {
 	}
 
 	cellCloseHandler(args: CellCloseEvent) {
@@ -100,47 +105,73 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 	}
 
 	AddRow() {
-		this.Cancel();
-		this.Mygrid.addRow(this.createFormGroup(this.newForm));
+		if (isNaN(this.rowIndex) || !this.CurrentSelectRow.dirty) {
+			this.Cancel("Add");
+			this.Mygrid.addRow(this.createFormGroup(this.newForm));
+		} else {
+			Swal.fire({
+				title: 'Please save the changes',
+				icon: 'warning',
+				showCloseButton: true,
+				cancelButtonText: 'Cancel',
+			})
+		}
 	}
 
-	DeleteRow() {
-		this.DataService.delete(this.dataItem[this.DataService.Key]).subscribe((res: any) => {
-			this.DataService.read(`$skip=${this.state.skip}&$top=10&$count=true`);
-			this.GetGridData();
-			this.alertService.success("Deleted Successfully");
-		});
+	async DeleteRow() {
+		await this.BeforeActionDelete();
+		this.StopDelete?.pipe(take(1)).subscribe((res: any) => {
+			if (res) {
+				this.DataService.delete(this.dataItem[this.DataService.Key]).subscribe((res: any) => {
+					this.DataService.read(`$skip=${this.state.skip}&$top=10&$count=true`);
+					this.alertService.success("Deleted Successfully");
+				});
+			}
+		})
 	};
 
 	public cellClickHandler(args: CellClickEvent): void {
 		this.dataItem = args.dataItem;
 		this.CellClick.emit(args);
-		if (!args.isEdited) {
+		if (isNaN(this.rowIndex) && !this.CurrentSelectRow.dirty) {
 			args.sender.editCell(
 				args.rowIndex,
 				args.columnIndex,
 				this.createFormGroup(args)
 			);
+		} else if (this.rowIndex == args.rowIndex || !this.CurrentSelectRow.dirty) {
+			args.sender.editCell(
+				args.rowIndex,
+				args.columnIndex,
+				this.createFormGroup(args)
+			);
+		} else {
+			Swal.fire({
+				title: 'Please save the changes',
+				icon: 'warning',
+				showCloseButton: true,
+				cancelButtonText: 'Cancel',
+			})
 		}
 	}
 
 	onValueChange(e: any) {
 		this.state.skip = e.skip;
 		this.DataService.read(toODataString(e) + '&$count=true');
-		this.GetGridData();
 	}
 
-	Cancel() {
+	Cancel(type?: string) {
 		this.DataService.read(`$skip=${this.state.skip}&$top=10&$count=true`);
-		this.GetGridData();
 		this.Mygrid.closeRow(this.rowIndex);
-		this.CurrentSelectRow.reset();
-		this.dataItem = null;
+		this.CurrentSelectRow.reset(type === "Add" ? {} : { ...this.dataItemReset });
+		this.dataItem = undefined;
+		this.dataItemReset = undefined;
+		this.rowIndex = undefined;
 	}
 
 	async Save() {
 		if (!this.dataItem) this.CreatedItemArray.push({ ...this.createFormGroup(this.newForm).value });
-		await this.BeforeAction();
+		await this.BeforeActionSave();
 		this.StopSave?.pipe(take(1)).subscribe((res: any) => {
 			if (res) {
 				if (isNaN(this.rowIndex)) {
@@ -149,7 +180,6 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 						this.DataService.add(this.CurrentSelectRow.getRawValue()).subscribe((res: any) => {
 							this.data['data'].unshift(this.CurrentSelectRow.getRawValue());
 							this.Mygrid.closeRow();
-							this.GetGridData();
 							this.handleFormGroup();
 							this.alertService.success("Saved Successfully");
 						});
@@ -160,7 +190,6 @@ export class BIGridComponent implements IGrid, OnInit, AfterViewInit {
 						this.DataService.edit(this.CurrentSelectRow.getRawValue(), this.CurrentSelectRow.getRawValue()[this.DataService.Key]).subscribe((res: any) => {
 							this.Mygrid.closeRow(this.rowIndex);
 							this.DataService.read(`$skip=${this.state.skip}&$top=10&$count=true`);
-							this.GetGridData();
 							this.handleFormGroup();
 							this.alertService.success("Saved Successfully");
 						});
